+++ packages/web/src/register.ts
@@ -3,3 +3,4 @@ import { defineWebGitGraph } from "./web-component";
 defineWebGitGraph();
 
 export { WebGitGraphElement, defineWebGitGraph } from "./web-component";
+export type { WebGitGraphElementEventMap } from "./events";